# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenOutlier API
  version: 1.0.0
  description: API for tracking channels, scanning niches, finding outliers, and saving references.
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: []
paths:
  /api/health:
    get:
      security: []
      summary: Health check
      responses:
        "200":
          description: OK
  /api/projects:
    get:
      summary: List projects
      responses:
        "200":
          description: Projects returned
    post:
      summary: Create project
      responses:
        "201":
          description: Project created
  /api/projects/{id}:
    get:
      summary: Get project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project returned
  /api/projects/{id}/source-sets:
    post:
      summary: Create source set
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Source set created
  /api/source-sets/{id}:
    get:
      summary: Get source set
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Source set returned
  /api/source-sets/{id}/channels:
    post:
      summary: Add channel to source set
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Channel attached
  /api/source-sets/{id}/discover:
    post:
      summary: Discover channels for source set
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suggestions returned
  /api/scan:
    post:
      summary: Trigger scan
      responses:
        "202":
          description: Scan started
  /api/scan/status:
    get:
      summary: Get scan status
      responses:
        "200":
          description: Scan status returned
  /api/discover/outliers:
    get:
      summary: Search outlier feed
      responses:
        "200":
          description: Outliers returned
  /api/discover/similar-topics:
    get:
      summary: Get similar topics for a video
      responses:
        "200":
          description: Similar topics returned
  /api/discover/niches:
    get:
      summary: Get niche clusters
      responses:
        "200":
          description: Niche data returned
  /api/projects/{id}/references/search:
    post:
      summary: Search references within a project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Search results returned
  /api/projects/{id}/references:
    get:
      summary: List saved references
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: References returned
    post:
      summary: Save reference
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Reference saved
  /api/projects/{id}/references/import-video:
    post:
      summary: Import one video as a reference
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Video imported